requirements.txt
asgiref==3.8.1
celery==5.5.3
Django==5.2.2
django-cors-headers==4.7.0
django-celery-beat==2.8.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
python-dotenv==1.1.0
redis==6.2.0
sqlparse==0.5.3
tzdata==2025.2
